LayoutTests/imported/w3c/web-platform-tests/inert/inert-canvas-fallback-content-expected.txt


PASS
      normal `div`

PASS
      normal `span`

PASS
      normal `a`

FAIL
      `div` in `display: none`
     assert_not_equals: got disallowed value Element node <div tabindex="-1" data-focusable="false">
      `div` in...
FAIL
      `span` in `display: none`
     assert_not_equals: got disallowed value Element node <span tabindex="-1" data-focusable="false">
      `span` ...
FAIL
      `a` in `display: none`
     assert_not_equals: got disallowed value Element node <a href="#" data-focusable="false">
      `a` in `display...
FAIL
      inert `div`
     assert_not_equals: got disallowed value Element node <div tabindex="-1" data-focusable="false">
      inert `d...
FAIL
      inert `span`
     assert_not_equals: got disallowed value Element node <span tabindex="-1" data-focusable="false">
      inert `...
FAIL
      inert `a`
     assert_not_equals: got disallowed value Element node <a href="#" data-focusable="false">
      inert `a`
    </a>
FAIL
      inert `div` in `display: none`
     assert_not_equals: got disallowed value Element node <div tabindex="-1" data-focusable="false">
      inert `d...
FAIL
      inert `span` in `display: none`
     assert_not_equals: got disallowed value Element node <span tabindex="-1" data-focusable="false">
      inert `...
FAIL
      inert `a` in `display: none`
     assert_not_equals: got disallowed value Element node <a href="#" data-focusable="false">
      inert `a` in `d...
